requirements.txt
openai>=1.50.0
pandas>=2.2.0
matplotlib>=3.8.0
numpy>=1.26.0
scipy>=1.11.0
python-dotenv>=1.0.1
tqdm>=4.66.0